agents/requirements.txt
strands-agents
boto3
httpx
bedrock-agentcore
